--- packages/core/src/core/cda_identity.py
@@ -1,112 +1,0 @@
-"""Compatibility facade for CDA identity, stable-key, and fallback helpers."""
-
-from core.cda_clinical_statement import (
-    CDA_CLINICAL_STATEMENT_LOCAL_NAMES,
-    CDA_CLINICAL_STATEMENT_TAGS,
-    CDA_SINGLE_STATEMENT_WRAPPER_LOCAL_NAMES,
-    _clinical_statement_for_identity,
-    _is_cda_clinical_statement,
-    _is_cda_single_clinical_statement_wrapper,
-    _is_hl7_element_with_local_name,
-    _single_direct_clinical_statement_child,
-)
-from core.cda_fallback_keys import (
-    _complete_root_extension_from_element,
-    _complete_root_extensions_from_direct_id_children,
-    _effective_time_discriminator,
-    _organizer_context,
-    _statement_code_pair,
-    _statement_effective_time,
-    _statement_id_root_extensions,
-    _weak_attribute_discriminator,
-    secondary_discriminator,
-    soft_context_key,
-)
-from core.cda_key_models import (
-    CodeKey,
-    DirectChildIdElementSetKey,
-    DirectChildTemplateIdElementSetKey,
-    DirectIdAttributeKey,
-    IdAttributeKeyBase,
-    NestedClinicalStatementIdAttributeKey,
-    NestedClinicalStatementIdElementSetKey,
-    NestedClinicalStatementTemplateIdElementSetKey,
-    NestedSectionIdElementSetKey,
-    NestedSectionTemplateIdElementSetKey,
-    RootExtension,
-    RootExtensionKey,
-    RootExtensionSetKeyBase,
-    StableKey,
-)
-from core.cda_narrative_keys import narrative_row_key, narrative_table_key
-from core.cda_stable_key import (
-    CODE_ATTRIBUTE,
-    CODE_SYSTEM_ATTRIBUTE,
-    DIRECT_ID_TAG,
-    DIRECT_TEMPLATE_ID_TAG,
-    ELEMENTS_HAVING_ROOT_EXTENSION_IDENTITY,
-    EXTENSION_ATTRIBUTE,
-    ROOT_ATTRIBUTE,
-    SECTION_TAG,
-    _attribute_key,
-    _code_key,
-    _direct_child_root_extensions_for_tag,
-    _id_attribute_key,
-    _nested_section_root_extensions_for_tag,
-    _root_extension_from_element,
-    _root_extension_key,
-    stable_key,
-)
-
-__all__ = [
-    "CDA_CLINICAL_STATEMENT_LOCAL_NAMES",
-    "CDA_CLINICAL_STATEMENT_TAGS",
-    "CDA_SINGLE_STATEMENT_WRAPPER_LOCAL_NAMES",
-    "CODE_ATTRIBUTE",
-    "CODE_SYSTEM_ATTRIBUTE",
-    "CodeKey",
-    "DIRECT_ID_TAG",
-    "DIRECT_TEMPLATE_ID_TAG",
-    "DirectChildIdElementSetKey",
-    "DirectChildTemplateIdElementSetKey",
-    "DirectIdAttributeKey",
-    "ELEMENTS_HAVING_ROOT_EXTENSION_IDENTITY",
-    "EXTENSION_ATTRIBUTE",
-    "IdAttributeKeyBase",
-    "NestedClinicalStatementIdAttributeKey",
-    "NestedClinicalStatementIdElementSetKey",
-    "NestedClinicalStatementTemplateIdElementSetKey",
-    "NestedSectionIdElementSetKey",
-    "NestedSectionTemplateIdElementSetKey",
-    "ROOT_ATTRIBUTE",
-    "RootExtension",
-    "RootExtensionKey",
-    "RootExtensionSetKeyBase",
-    "SECTION_TAG",
-    "StableKey",
-    "_attribute_key",
-    "_clinical_statement_for_identity",
-    "_code_key",
-    "_complete_root_extension_from_element",
-    "_complete_root_extensions_from_direct_id_children",
-    "_direct_child_root_extensions_for_tag",
-    "_effective_time_discriminator",
-    "_id_attribute_key",
-    "_is_cda_clinical_statement",
-    "_is_cda_single_clinical_statement_wrapper",
-    "_is_hl7_element_with_local_name",
-    "_nested_section_root_extensions_for_tag",
-    "_organizer_context",
-    "_root_extension_from_element",
-    "_root_extension_key",
-    "_single_direct_clinical_statement_child",
-    "_statement_code_pair",
-    "_statement_effective_time",
-    "_statement_id_root_extensions",
-    "_weak_attribute_discriminator",
-    "narrative_row_key",
-    "narrative_table_key",
-    "secondary_discriminator",
-    "soft_context_key",
-    "stable_key",
-]
