+++ packages/core/test/unit/test_cda_stable_key.py
@@ -1,10 +1,8 @@
-from core.cda_clinical_statement import _clinical_statement_for_identity
 from core.cda_fallback_keys import secondary_discriminator, soft_context_key
 from core.cda_key_models import (
     CodeKey,
     DirectChildIdElementSetKey,
     DirectChildTemplateIdElementSetKey,
-    DirectIdAttributeKey,
     NestedClinicalStatementIdAttributeKey,
     NestedClinicalStatementIdElementSetKey,
     NestedClinicalStatementTemplateIdElementSetKey,
@@ -19,26 +17,6 @@ from core.cda_stable_key import (
     stable_key,
 )
 from helpers import HL7_NS, elem, observation
-
-
-def test_id_attribute_key_classes_are_distinct_identity_variants():
-    assert DirectIdAttributeKey(name="ID", value="same-id") != (
-        NestedClinicalStatementIdAttributeKey(name="ID", value="same-id")
-    )
-
-
-def test_root_extension_set_key_classes_are_distinct_identity_variants():
-    root_extensions = (RootExtension(root="same-root", extension="same-extension"),)
-
-    assert DirectChildIdElementSetKey(root_extensions=root_extensions) != (
-        NestedClinicalStatementIdElementSetKey(root_extensions=root_extensions)
-    )
-    assert DirectChildIdElementSetKey(root_extensions=root_extensions) != (
-        NestedSectionIdElementSetKey(root_extensions=root_extensions)
-    )
-    assert DirectChildTemplateIdElementSetKey(root_extensions=root_extensions) != (
-        DirectChildIdElementSetKey(root_extensions=root_extensions)
-    )
 
 
 def test_direct_child_root_extensions_for_tag_sorts_deduplicates_and_skips_missing_root():
@@ -343,83 +321,6 @@ def test_unrelated_descendant_id_is_not_stable_identity():
     )
 
     assert stable_key(wrapper) is None
-
-
-def test_clinical_statement_identity_unwraps_single_statement_wrappers():
-    for wrapper_name in ("entry", "entryRelationship", "component"):
-        wrapper = elem(
-            f"""
-            <{wrapper_name} xmlns="{HL7_NS}">
-              <observation classCode="OBS" moodCode="EVN"/>
-            </{wrapper_name}>
-            """
-        )
-        observation_element = wrapper.xpath(
-            "./hl7:observation",
-            namespaces={"hl7": HL7_NS},
-        )[0]
-
-        assert _clinical_statement_for_identity(wrapper) is observation_element
-
-
-def test_clinical_statement_identity_does_not_unwrap_multi_entry_container():
-    section = elem(
-        f"""
-        <section xmlns="{HL7_NS}">
-          <entry>
-            <observation classCode="OBS" moodCode="EVN">
-              <id root="first"/>
-            </observation>
-          </entry>
-          <entry>
-            <observation classCode="OBS" moodCode="EVN">
-              <id root="second"/>
-            </observation>
-          </entry>
-        </section>
-        """
-    )
-
-    assert _clinical_statement_for_identity(section) is None
-    assert stable_key(section) is None
-
-
-def test_single_statement_wrapper_with_multiple_direct_statements_has_no_statement_identity():
-    entry = elem(
-        f"""
-        <entry xmlns="{HL7_NS}">
-          <observation classCode="OBS" moodCode="EVN">
-            <id root="first"/>
-          </observation>
-          <observation classCode="OBS" moodCode="EVN">
-            <id root="second"/>
-          </observation>
-        </entry>
-        """
-    )
-
-    assert _clinical_statement_for_identity(entry) is None
-    assert stable_key(entry) is None
-
-
-def test_organizer_uses_itself_for_clinical_statement_identity():
-    organizer = elem(
-        f"""
-        <organizer xmlns="{HL7_NS}" classCode="BATTERY" moodCode="EVN">
-          <id root="organizer-id" extension="1"/>
-          <component>
-            <observation classCode="OBS" moodCode="EVN">
-              <id root="observation-id" extension="1"/>
-            </observation>
-          </component>
-        </organizer>
-        """
-    )
-
-    assert _clinical_statement_for_identity(organizer) is organizer
-    assert stable_key(organizer) == DirectChildIdElementSetKey(
-        root_extensions=(RootExtension(root="organizer-id", extension="1"),),
-    )
 
 
 def test_soft_context_key_uses_full_direct_template_id_root_extensions():
